Moved classes into packages.
